+++ dlt/__init__.py
@@ -30,8 +30,7 @@ from dlt import sources
 from dlt.extract.decorators import source, resource, transformer, defer
 from dlt.destinations.decorators import destination
 from dlt.transformations.decorators import transformation
-from dlt.common.destination.dataset import Dataset, Relation
-from dlt.destinations.dataset import dataset
+from dlt._dataset import dataset, Dataset, Relation
 
 from dlt.pipeline import (
     pipeline as _pipeline,
