+++ dlt/_dataset/__init__.py
@@ -1,0 +1,8 @@
+from dlt._dataset.dataset import Dataset, dataset
+from dlt._dataset.relation import Relation
+
+__all__ = (
+    "dataset",
+    "Dataset",
+    "Relation",
+)
