+++ dlt/destinations/dataset/__init__.py
@@ -1,7 +1,17 @@
+from typing import TYPE_CHECKING
+
+if not TYPE_CHECKING:
+    from dlt.common.warnings import DltDeprecationWarning
+
+    DltDeprecationWarning(
+        "Content from this module was moved to `dlt._dataset`, which is internal.",
+        since="1.15",
+        expected_due="2.0",
+    )
+
 from dlt.destinations.dataset.factory import dataset
-from dlt.destinations.dataset.dataset import (
-    ReadableDBAPIDataset,
-)
+from dlt.destinations.dataset.dataset import ReadableDBAPIDataset
+from dlt.destinations.dataset.relation import ReadableDBAPIRelation
 
 from dlt.destinations.dataset.utils import (
     get_destination_clients,
@@ -12,7 +22,7 @@ from dlt.destinations.dataset.utils import (
 __all__ = [
     "dataset",
     "ReadableDBAPIDataset",
-    "BaseReadableDBAPIDataset",
+    "ReadableDBAPIRelation",
     "get_destination_client_initial_config",
     "get_destination_clients",
 ]
