+++ dlt/destinations/dataset/exceptions.py
@@ -1,22 +1,20 @@
-from dlt.common.exceptions import DltException
+from typing import TYPE_CHECKING
 
+if not TYPE_CHECKING:
+    from dlt.common.warnings import DltDeprecationWarning
 
-class DatasetException(DltException):
-    pass
+    DltDeprecationWarning(
+        """Content from this module was moved to `dlt._dataset.exceptions`, which is internal. \
+        You can catch dlt exceptions using `dlt.common.exceptions.DltException.""",
+        since="1.15",
+        expected_due="2.0",
+    )
 
+from dlt._dataset.exceptions import (
+    DatasetException,
+    RelationHasQueryException,
+    RelationUnknownColumnException,
+)
 
-class ReadableRelationHasQueryException(DatasetException):
-    def __init__(self, attempted_change: str) -> None:
-        msg = (
-            "This readable relation was created with a provided sql query. You cannot change"
-            f" `{attempted_change}`. Please change the orignal sql query."
-        )
-        super().__init__(msg)
-
-
-class ReadableRelationUnknownColumnException(DatasetException):
-    def __init__(self, column_name: str) -> None:
-        msg = (
-            f"The selected column `{column_name}` is not known in the dlt schema for this relation."
-        )
-        super().__init__(msg)
+ReadableRelationHasQueryException = RelationHasQueryException
+ReadableRelationUnknownColumnException = RelationUnknownColumnException
