+++ dlt/helpers/dashboard/utils.py
@@ -91,7 +91,7 @@ def get_pipeline(pipeline_name: str, pipelines_dir: str) -> dlt.Pipeline:
 def get_destination_config(pipeline: dlt.Pipeline) -> DestinationClientConfiguration:
     """Get the destination config of a pipeline."""
     # NOTE: this uses internal interfaces for now...
-    return cast(DestinationClientConfiguration, pipeline.dataset().destination_client.config)  # type: ignore[attr-defined]
+    return pipeline.dataset().destination_client.config
 
 
 def pipeline_details(pipeline: dlt.Pipeline) -> List[Dict[str, Any]]:
