+++ tests/load/test_read_interfaces.py
@@ -20,7 +20,6 @@ from dlt.common.schema.typing import TTableFormat
 from dlt.extract.source import DltSource
 from dlt.destinations.dataset import dataset as _dataset
 from dlt.transformations.exceptions import LineageFailedException
-from dlt.destinations.dataset.dataset import ReadableDBAPIDataset, ReadableDBAPIRelation
 
 from tests.load.utils import (
     destinations_configs,
@@ -192,15 +191,13 @@ def test_str_and_repr_on_dataset_and_relation(populated_pipeline: Pipeline) -> N
     if populated_pipeline.destination.destination_type != "dlt.destinations.duckdb":
         pytest.skip("Only duckdb is supported for this test")
 
-    dataset_ = cast(ReadableDBAPIDataset, populated_pipeline.dataset())
+    dataset_ = populated_pipeline.dataset()
 
     def _replace_variable_content(s: str) -> str:
         # replace dataset name
         s = s.replace(dataset_.dataset_name, "dataset_name")
         # replace destination config
-        dest_config = str(
-            cast(ReadableDBAPIDataset, populated_pipeline.dataset()).destination_client.config
-        )
+        dest_config = str(populated_pipeline.dataset().destination_client.config)
         s = s.replace(dest_config, "<destination_config>")
         return s
 
@@ -525,7 +522,7 @@ def test_sql_queries(populated_pipeline: Pipeline) -> None:
         "select * from items where id < 20"
     )
 
-    assert query_relationship._sqlglot_expression == query_from_query_function._sqlglot_expression  # type: ignore
+    assert query_relationship._sqlglot_expression == query_from_query_function._sqlglot_expression
 
     # we selected the first 20
     table = query_relationship.arrow()
@@ -549,7 +546,7 @@ def test_sql_queries(populated_pipeline: Pipeline) -> None:
         " as di ON (i.id = di.id) WHERE i.id < 20 ORDER BY i.id ASC"
     )
 
-    join_relationship = cast(ReadableDBAPIRelation, populated_pipeline.dataset()(query))
+    join_relationship = populated_pipeline.dataset()(query)
     table = join_relationship.fetchall()
     assert len(table) == 20
 
@@ -601,7 +598,7 @@ def test_sql_queries(populated_pipeline: Pipeline) -> None:
     ids=lambda x: x.name,
 )
 def test_limit_and_head(populated_pipeline: Pipeline) -> None:
-    dataset_ = cast(ReadableDBAPIDataset, populated_pipeline.dataset())
+    dataset_ = populated_pipeline.dataset()
 
     # test sql_client lifecycle
     assert dataset_._opened_sql_client is None
@@ -627,7 +624,7 @@ def test_limit_and_head(populated_pipeline: Pipeline) -> None:
     assert table_relationship.limit(24).arrow().num_rows == 24
     assert dataset_._opened_sql_client is None
 
-    limit_relationship = cast(ReadableDBAPIRelation, table_relationship.limit(24))
+    limit_relationship = table_relationship.limit(24)
     for data_ in limit_relationship.iter_fetch(6):
         assert len(data_) == 6
         # client stays open
@@ -635,7 +632,7 @@ def test_limit_and_head(populated_pipeline: Pipeline) -> None:
 
     # run multiple requests on one connection
     with dataset_ as d_:
-        limit_relationship = cast(ReadableDBAPIRelation, table_relationship.limit(24))
+        limit_relationship = table_relationship.limit(24)
         for _data in limit_relationship.iter_fetch(6):
             # client stays open
             assert limit_relationship._opened_sql_client is not None
@@ -644,7 +641,7 @@ def test_limit_and_head(populated_pipeline: Pipeline) -> None:
                 == d_._opened_sql_client.native_connection
             )
 
-        other_relationship = cast(ReadableDBAPIRelation, table_relationship.limit(10))
+        other_relationship = table_relationship.limit(10)
         for _data in other_relationship.iter_fetch(6):
             assert other_relationship._opened_sql_client is not None
             assert (
@@ -670,7 +667,7 @@ def test_limit_and_head(populated_pipeline: Pipeline) -> None:
 )
 def test_dataset_client_caching_and_connection_handling(populated_pipeline: Pipeline) -> None:
     # no clients exist yet
-    dataset = cast(ReadableDBAPIDataset, populated_pipeline.dataset())
+    dataset = populated_pipeline.dataset()
     assert dataset._opened_sql_client is None
     assert dataset._sql_client is None
 
@@ -881,7 +878,7 @@ def test_where(populated_pipeline: Pipeline) -> None:
     assert total_records - 3 == len(not_in_rows)
 
     with pytest.raises(ValueErrorWithKnownValues) as py_exc:
-        not_in_rows = items.filter("id", "wrong", [0, 1, 2]).fetchall()  # type: ignore
+        not_in_rows = items.filter("id", "wrong", [0, 1, 2]).fetchall()
 
     assert "Received invalid value `operator=wrong`" in py_exc.value.args[0]
 
@@ -1064,8 +1061,8 @@ def test_ibis_expression_relation(populated_pipeline: Pipeline) -> None:
     # verify computed columns
     assert (
         list(relation.columns_schema.keys())
-        == relation.columns  # type: ignore[attr-defined]
-        == relation._ipython_key_completions_()  # type: ignore[attr-defined]
+        == relation.columns
+        == relation._ipython_key_completions_()
         == ["id", "double_id"]
     )
 
@@ -1075,8 +1072,8 @@ def test_ibis_expression_relation(populated_pipeline: Pipeline) -> None:
     assert agg_relation.fetchone()[0] == reduce(lambda a, b: a + b, range(20))
     assert (
         list(agg_relation.columns_schema.keys())
-        == agg_relation.columns  # type: ignore[attr-defined]
-        == agg_relation._ipython_key_completions_()  # type: ignore[attr-defined]
+        == agg_relation.columns
+        == agg_relation._ipython_key_completions_()
         == ["sum_id"]
     )
 
@@ -1091,7 +1088,7 @@ def test_ibis_expression_relation(populated_pipeline: Pipeline) -> None:
     # also we return the keys of the discovered schema columns
     def sql_from_expr(expr: Any) -> Tuple[str, List[str]]:
         relation = dataset(expr)
-        query = str(relation.to_sql()).replace(populated_pipeline.dataset_name, "dataset")  # type: ignore[attr-defined]
+        query = str(relation.to_sql()).replace(populated_pipeline.dataset_name, "dataset")
         columns = list(relation.columns_schema.keys()) if relation.columns_schema else None
         return re.sub(r"\s+", " ", query), columns
 
@@ -1358,14 +1355,11 @@ def test_standalone_dataset(populated_pipeline: Pipeline) -> None:
     total_records = _total_records(populated_pipeline.destination.destination_type)
 
     # check dataset factory
-    dataset = cast(
-        ReadableDBAPIDataset,
-        _dataset(
-            destination=populated_pipeline.destination,
-            dataset_name=populated_pipeline.dataset_name,
-            # use name otherwise aleph schema is loaded
-            schema=populated_pipeline.default_schema_name,
-        ),
+    dataset = dlt.dataset(
+        destination=populated_pipeline.destination,
+        name=populated_pipeline.dataset_name,
+        # use name otherwise aleph schema is loaded
+        schema=populated_pipeline.default_schema_name,
     )
     # verify that sql client and schema are lazy loaded
     assert not dataset._schema
@@ -1376,24 +1370,18 @@ def test_standalone_dataset(populated_pipeline: Pipeline) -> None:
     assert dataset.schema.tables["items"]["write_disposition"] == "replace"
 
     # check that schema is not loaded when wrong name given
-    dataset = cast(
-        ReadableDBAPIDataset,
-        _dataset(
-            destination=populated_pipeline.destination,
-            dataset_name=populated_pipeline.dataset_name,
-            schema="wrong_schema_name",
-        ),
+    dataset = dlt.dataset(
+        destination=populated_pipeline.destination,
+        name=populated_pipeline.dataset_name,
+        schema="wrong_schema_name",
     )
     assert "items" not in dataset.schema.tables
     assert dataset.schema.name == "wrong_schema_name"
 
     # check that schema is loaded if no schema name given
-    dataset = cast(
-        ReadableDBAPIDataset,
-        _dataset(
-            destination=populated_pipeline.destination,
-            dataset_name=populated_pipeline.dataset_name,
-        ),
+    dataset = dlt.dataset(
+        destination=populated_pipeline.destination,
+        name=populated_pipeline.dataset_name,
     )
     # aleph is a secondary schema in the pipeline but because it was stored second
     # will be retrieved by default
@@ -1401,12 +1389,9 @@ def test_standalone_dataset(populated_pipeline: Pipeline) -> None:
     assert dataset.schema.tables["digits"]["write_disposition"] == "append"
 
     # check that there is no error when creating dataset without schema table
-    dataset = cast(
-        ReadableDBAPIDataset,
-        _dataset(
-            destination=populated_pipeline.destination,
-            dataset_name="unknown_dataset",
-        ),
+    dataset = dlt.dataset(
+        destination=populated_pipeline.destination,
+        name="unknown_dataset",
     )
     assert dataset.schema.name == "unknown_dataset"
     assert "items" not in dataset.schema.tables
@@ -1424,12 +1409,9 @@ def test_standalone_dataset(populated_pipeline: Pipeline) -> None:
     with populated_pipeline.destination_client() as client:
         client.update_stored_schema()
 
-    dataset = cast(
-        ReadableDBAPIDataset,
-        _dataset(
-            destination=populated_pipeline.destination,
-            dataset_name=populated_pipeline.dataset_name,
-        ),
+    dataset = dlt.dataset(
+        destination=populated_pipeline.destination,
+        name=populated_pipeline.dataset_name,
     )
     assert dataset.schema.name == "some_other_schema"
     assert "other_table" in dataset.schema.tables
@@ -1514,7 +1496,7 @@ def test_naming_convention_propagation(destination_config: DestinationTestConfig
     dataset_ = pipeline.dataset()
     df = dataset_.ItemS.df()
     assert df.columns.tolist()[0] == "ID"
-    with dataset_.sql_client as client:  # type: ignore
+    with dataset_.sql_client as client:
         assert client.dataset_name.startswith("Read_test")
         tables = client.native_connection.sql("SHOW TABLES;")
         assert "ItemS" in str(tables)
