+++ tests/transformations/test_transformation_decorator.py
@@ -9,15 +9,12 @@ from dlt.common import logger
 
 from dlt.common.configuration.inject import get_fun_last_config, get_fun_spec
 from dlt.common.configuration.specs.base_configuration import configspec
-from dlt.common.destination.dataset import Dataset
 from dlt.common.schema.schema import Schema
 from dlt.common.schema.utils import new_table
-from dlt.extract.hints import SqlModel
 from dlt.transformations.configuration import TransformationConfiguration
 from dlt.transformations.exceptions import IncompatibleDatasetsException, TransformationException
 from dlt.extract.exceptions import ResourceExtractionError
-from dlt.common.destination.dataset import Relation
-from dlt.destinations.dataset.dataset import ReadableDBAPIRelation
+from dlt.destinations.dataset.relation import Relation
 
 
 def test_no_datasets_used() -> None:
@@ -46,7 +43,7 @@ def test_no_datasets_used() -> None:
 def test_iterator_function_as_transform_function() -> None:
     # test that a generator function is used as a regular resource
     @dlt.transformation()
-    def transform(dataset: Dataset) -> Any:
+    def transform(dataset: dlt.Dataset) -> Any:
         yield [{"some": "data"}]
 
     assert list(transform(dlt.dataset("duckdb", "dataset_name"))) == [{"some": "data"}]
@@ -61,7 +58,7 @@ def test_incremental_argument_is_not_supported(caplog: LogCaptureFixture) -> Non
 
             @dlt.transformation()
             def transform_1(
-                dataset: Dataset,
+                dataset: dlt.Dataset,
                 incremental_arg=dlt.sources.incremental("col1"),
             ) -> Any:
                 yield "SELECT col1 FROM table1"
@@ -81,7 +78,7 @@ def test_incremental_argument_is_not_supported(caplog: LogCaptureFixture) -> Non
 
             @dlt.transformation()
             def transform_2(
-                dataset: Dataset,
+                dataset: dlt.Dataset,
                 # TODO: this may be edge case when we have native value but nevertheless dlt complains that there's no default
                 incremental_arg: Optional[dlt.sources.incremental] = "some_value",  # type: ignore
             ) -> Any:
@@ -101,7 +98,7 @@ def test_incremental_argument_is_not_supported(caplog: LogCaptureFixture) -> Non
         with pytest.raises(ResourceExtractionError):
 
             @dlt.transformation()
-            def transform_3(dataset: Dataset) -> Any:
+            def transform_3(dataset: dlt.Dataset) -> Any:
                 return "SELECT col1 FROM table1"
 
             list(transform_3(dlt.dataset("duckdb", "dataset_name")))
@@ -132,7 +129,7 @@ def test_base_transformation_spec() -> None:
 
     # we return SQL so we expect a model to be created
     model = list(default_spec(ds_))[0]
-    assert isinstance(model, ReadableDBAPIRelation)
+    assert isinstance(model, Relation)
     # TODO: why dialect is not set??
     # assert model.dialect is not None
 
@@ -152,7 +149,7 @@ def test_base_transformation_spec() -> None:
     os.environ["LAST_ID"] = "test_last_id"
 
     model = list(default_transformation_with_args(ds_))[0]
-    assert isinstance(model, ReadableDBAPIRelation)
+    assert isinstance(model, Relation)
     assert get_fun_last_config(default_transformation_with_args)["last_id"] == "test_last_id"
 
     # test explicit spec
@@ -184,7 +181,7 @@ def test_base_transformation_spec() -> None:
     os.environ["SOURCES__DEFAULT_NAME_OVR__LIMIT"] = "100"
 
     model = list(default_transformation_spec(ds_))[0]
-    assert isinstance(model, ReadableDBAPIRelation)
+    assert isinstance(model, Relation)
     query = model.to_sql()
     # make sure we have our args in query
     assert "uniq_last_id" in query
